files/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Audio processing
pydub==0.25.1

# ── Local MusicGen ────────────────────────────────────────────────────────────
# audiocraft on PyPI pins torch==2.1.0 which has NO Python 3.12 wheels.
# Do NOT add audiocraft/torch here — install them manually in order (see README).